+++ bcbio/distributed/ipythontasks.py
@@ -290,6 +290,12 @@ def detect_sv(*args):
         return ipython.zip_args(apply(structural.detect_sv, *args))
 
 @require(structural)
+def validate_sv(*args):
+    args = ipython.unzip_args(args)
+    with _setup_logging(args) as config:
+        return ipython.zip_args(apply(structural.validate_sv, *args))
+
+@require(structural)
 def finalize_sv(*args):
     args = ipython.unzip_args(args)
     with _setup_logging(args) as config: