+++ bcbio/structural/__init__.py
@@ -7,18 +7,18 @@ import operator
 import toolz as tz
 
 from bcbio.pipeline import datadict as dd
-from bcbio.structural import (battenberg, cn_mops, cnvkit, delly, ensemble,
+from bcbio.structural import (battenberg, cn_mops, cnvkit, delly,
                               lumpy, manta, metasv, plot, validate, wham)
 from bcbio.variation import vcfutils
 
-_CALLERS = {}
-_SOMATIC_CALLERS = {"manta": manta.run}
-_BATCH_CALLERS = {"cn.mops": cn_mops.run, "cnvkit": cnvkit.run,
-                  "delly": delly.run, "lumpy": lumpy.run, "wham": wham.run,
-                  "battenberg": battenberg.run}
-_ENSEMBLE_CALLERS = {"metasv": metasv.run}
+# Stratify callers by stage -- see `run` documentation below for definitions
+_CALLERS = {
+  "initial": {"cnvkit": cnvkit.run,
+              "battenberg": battenberg.run},
+  "standard": {"cn.mops": cn_mops.run, "manta": manta.run,
+               "delly": delly.run, "lumpy": lumpy.run, "wham": wham.run},
+  "ensemble": {"metasv": metasv.run}}
 _NEEDS_BACKGROUND = set(["cn.mops"])
-_INITIAL_CALLERS = set(["battenberg", "cnvkit"])
 
 def _get_svcallers(data):
     svs = data["config"]["algorithm"].get("svcaller")
@@ -28,19 +28,19 @@ def _get_svcallers(data):
         svs = [svs]
     return svs
 
-def _handle_multiple_svcallers(data):
+def _handle_multiple_svcallers(data, stage):
     """Retrieve configured structural variation caller, handling multiple.
     """
     svs = _get_svcallers(data)
     out = []
     for svcaller in svs:
-        if svcaller not in _ENSEMBLE_CALLERS:
+        if svcaller in _CALLERS[stage]:
             base = copy.deepcopy(data)
             base["config"]["algorithm"]["svcaller_active"] = svcaller
             out.append(base)
     return out
 
-def finalize_sv(samples, config, initial_only=False):
+def finalize_sv(samples, config):
     """Combine results from multiple sv callers into a single ordered 'sv' key.
 
     Handles ensemble calling and plotting of results.
@@ -60,13 +60,7 @@ def finalize_sv(samples, config, initial_only=False):
                                 key=orig_svcaller_order)
         final = grouped_calls[0]
         if len(sorted_svcalls) > 0:
-            final_calls = reduce(operator.add, [x["sv"] for x in sorted_svcalls])
-            if not initial_only:
-                for caller in (c for c in _get_svcallers(final) if c in _ENSEMBLE_CALLERS):
-                    final_calls = _ENSEMBLE_CALLERS[caller](final_calls, final)
-                final_calls = ensemble.summarize(final_calls, final, grouped_calls)
-                final_calls = validate.evaluate(final, final_calls)
-            final["sv"] = final_calls
+            final["sv"] = reduce(operator.add, [x["sv"] for x in sorted_svcalls])
         del final["config"]["algorithm"]["svcaller_active"]
         batch = dd.get_batch(final) or dd.get_sample_name(final)
         batches = batch if isinstance(batch, (list, tuple)) else [batch]
@@ -87,68 +81,62 @@ def finalize_sv(samples, config, initial_only=False):
                 out.append([data])
     return out
 
-def run(samples, run_parallel, initial_only=False):
+def validate_sv(data):
+    """Validate structural variant calls for a sample.
+    """
+    return [[validate.evaluate(data)]]
+
+def run(samples, run_parallel, stage):
     """Run structural variation detection.
 
-    initial_only indicates we should run structural variation inputs, like
-    CNV calling, we can use to inform low frequency variant calling.
+    The stage indicates which level of structural variant calling to run.
+      - initial, run prior to other callers and variant calling
+      - standard, regular batch calling
+      - ensemble, post-calling, combine other callers-
     """
     to_process = collections.OrderedDict()
     extras = []
     background = []
     for data in (xs[0] for xs in samples):
-        ready_data = _handle_multiple_svcallers(data)
+        ready_data = _handle_multiple_svcallers(data, stage)
         if len(ready_data) > 0:
             background.append(data)
             for x in ready_data:
                 svcaller = x["config"]["algorithm"].get("svcaller_active")
-                # reset SV information if we're running a second pass SV call
-                if "sv" in x:
-                    del x["sv"]
-                batch = dd.get_batch(x)
-                paired = vcfutils.get_paired_phenotype(x)
-                if ((svcaller in _BATCH_CALLERS and batch) or
-                      (svcaller in _SOMATIC_CALLERS and paired and batch)):
-                    batches = batch if isinstance(batch, (list, tuple)) else [batch]
-                    for b in batches:
-                        try:
-                            to_process[(svcaller, b)].append(x)
-                        except KeyError:
-                            to_process[(svcaller, b)] = [x]
+                if stage == "ensemble":  # no batching for ensemble methods
+                    batch = dd.get_sample_name(x)
                 else:
-                    to_process[(svcaller, dd.get_sample_name(x))] = [x]
+                    batch = dd.get_batch(x) or dd.get_sample_name(x)
+                batches = batch if isinstance(batch, (list, tuple)) else [batch]
+                for b in batches:
+                    try:
+                        to_process[(svcaller, b)].append(x)
+                    except KeyError:
+                        to_process[(svcaller, b)] = [x]
         else:
             extras.append([data])
-    processed = run_parallel("detect_sv", ([xs, background, xs[0]["config"], initial_only]
+    processed = run_parallel("detect_sv", ([xs, background, xs[0]["config"], stage]
                                            for xs in to_process.values()))
-    finalized = (run_parallel("finalize_sv", [([xs[0] for xs in processed], processed[0][0]["config"],
-                                               initial_only)])
+    finalized = (run_parallel("finalize_sv", [([xs[0] for xs in processed], processed[0][0]["config"])])
                  if len(processed) > 0 else [])
     return extras + finalized
 
-def detect_sv(items, all_items, config, initial_only=False):
+def detect_sv(items, all_items, config, stage):
     """Top level parallel target for examining structural variation.
     """
     svcaller = config["algorithm"].get("svcaller_active")
+    caller_fn = _CALLERS[stage].get(svcaller)
     out = []
-    if svcaller and (not initial_only or svcaller in _INITIAL_CALLERS):
-        if svcaller in _CALLERS:
-            assert len(items) == 1
-            data = items[0]
-            data["sv"] = _CALLERS[svcaller](data)
-            out.append([data])
-        elif svcaller in _BATCH_CALLERS or svcaller in _SOMATIC_CALLERS:
-            if (svcaller in _NEEDS_BACKGROUND and
-                  not vcfutils.is_paired_analysis([x.get("align_bam") for x in items], items)):
-                names = set([tz.get_in(["rgnames", "sample"], x) for x in items])
-                background = [x for x in all_items if tz.get_in(["rgnames", "sample"], x) not in names]
-                for svdata in _BATCH_CALLERS[svcaller](items, background):
-                    out.append([svdata])
-            else:
-                for svdata in _BATCH_CALLERS.get(svcaller, _SOMATIC_CALLERS.get(svcaller))(items):
-                    out.append([svdata])
+    if svcaller and caller_fn:
+        if (svcaller in _NEEDS_BACKGROUND and
+                not vcfutils.is_paired_analysis([x.get("align_bam") for x in items], items)):
+            names = set([tz.get_in(["rgnames", "sample"], x) for x in items])
+            background = [x for x in all_items if tz.get_in(["rgnames", "sample"], x) not in names]
+            for svdata in caller_fn(items, background):
+                out.append([svdata])
         else:
-            raise ValueError("Unexpected structural variant caller: %s" % svcaller)
+            for svdata in caller_fn(items):
+                out.append([svdata])
     else:
         for data in items:
             out.append([data])